split animation logic from scroll calc
