+++ src/index.ts
@@ -1,4 +1,5 @@
 import animate from 'amator'
+import calculate, { Options as CalculateOptions } from './calculate'
 
 // Legacy
 export interface AnimateOptions {
@@ -21,14 +22,27 @@ export interface Offset {
   left?: number
 }
 
-export interface Options {
-  boundary?: Element
-  centerIfNeeded?: boolean
+const handleScroll = (parent, { scrollLeft, scrollTop }, config) => {
+  if (config.duration) {
+    animate(
+      parent,
+      {
+        scrollLeft: scrollLeft,
+        scrollTop: scrollTop,
+      },
+      { duration: config.duration, easing: config.easing }
+    )
+  } else {
+    parent.scrollLeft = scrollLeft
+    parent.scrollTop = scrollTop
+  }
+}
+
+export interface Options extends CalculateOptions {
   // Setting a duration will enable animations
   duration?: number
   // Easing only take effect if a duration is set
   easing?: 'ease' | 'easeIn' | 'easeOut' | 'easeInOut' | 'linear'
-  offset?: Offset
 }
 
 function isBoolean(options: boolean | Options): options is boolean {
@@ -45,7 +59,7 @@ export default function scrollIntoViewIfNeeded(
   if (!target || !(target instanceof HTMLElement))
     throw new Error('Element is required in scrollIntoViewIfNeeded')
 
-  let config: Options = { centerIfNeeded: false }
+  let config: Options = { centerIfNeeded: false, handleScroll }
 
   if (isBoolean(options)) {
     config.centerIfNeeded = options
@@ -80,110 +94,5 @@ export default function scrollIntoViewIfNeeded(
     config.offset.left = offsetOptions.offsetLeft
   }
 
-  function withinBounds(value, min, max, extent) {
-    if (
-      config.centerIfNeeded === false ||
-      (max <= value + extent && value <= min + extent)
-    ) {
-      return Math.min(max, Math.max(min, value))
-    } else {
-      return (min + max) / 2
-    }
-  }
-
-  const { offset } = config
-  const offsetTop = offset.top
-  const offsetLeft = offset.left
-  const offsetBottom = offset.bottom
-  const offsetRight = offset.right
-
-  function makeArea(left, top, width, height) {
-    return {
-      left: left + offsetLeft,
-      top: top + offsetTop,
-      width: width,
-      height: height,
-      right: left + offsetLeft + width + offsetRight,
-      bottom: top + offsetTop + height + offsetBottom,
-      translate: function(x, y) {
-        return makeArea(
-          x + left + offsetLeft,
-          y + top + offsetTop,
-          width,
-          height
-        )
-      },
-      relativeFromTo: function(lhs, rhs) {
-        let newLeft = left + offsetLeft,
-          newTop = top + offsetTop
-        lhs = lhs.offsetParent
-        rhs = rhs.offsetParent
-        if (lhs === rhs) {
-          return area
-        }
-        for (; lhs; lhs = lhs.offsetParent) {
-          newLeft += lhs.offsetLeft + lhs.clientLeft
-          newTop += lhs.offsetTop + lhs.clientTop
-        }
-        for (; rhs; rhs = rhs.offsetParent) {
-          newLeft -= rhs.offsetLeft + rhs.clientLeft
-          newTop -= rhs.offsetTop + rhs.clientTop
-        }
-        return makeArea(newLeft, newTop, width, height)
-      },
-    }
-  }
-
-  let parent,
-    area = makeArea(
-      target.offsetLeft,
-      target.offsetTop,
-      target.offsetWidth,
-      target.offsetHeight
-    )
-  while (
-    (parent = target.parentNode) instanceof HTMLElement &&
-    target !== config.boundary
-  ) {
-    const clientLeft = parent.offsetLeft + parent.clientLeft
-    const clientTop = parent.offsetTop + parent.clientTop
-
-    // Make area relative to parent's client area.
-    area = area
-      .relativeFromTo(target, parent)
-      .translate(-clientLeft, -clientTop)
-
-    const scrollLeft = withinBounds(
-      parent.scrollLeft,
-      area.right - parent.clientWidth,
-      area.left,
-      parent.clientWidth
-    )
-    const scrollTop = withinBounds(
-      parent.scrollTop,
-      area.bottom - parent.clientHeight,
-      area.top,
-      parent.clientHeight
-    )
-    if (config.duration) {
-      animate(
-        parent,
-        {
-          scrollLeft: scrollLeft,
-          scrollTop: scrollTop,
-        },
-        { duration: config.duration, easing: config.easing }
-      )
-    } else {
-      parent.scrollLeft = scrollLeft
-      parent.scrollTop = scrollTop
-    }
-
-    // Determine actual scroll amount by reading back scroll properties.
-    area = area.translate(
-      clientLeft - parent.scrollLeft,
-      clientTop - parent.scrollTop
-    )
-    target = parent
-  }
+  return calculate(target, config)
 }
